Fix Filtering list ports
Currently the Filtering list is forwarding the ports to the ports of the wrapped list but this is wrong.
The Filtering list needs to go itself into the #items: method, else we cannot save the items to filter. 
Currently if you use transmissions on a filtering list, the list ends up empty because the unflitered items list will be an empty array